Fix auto-restore rounding error
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@314 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/HPRestoreItemList.java
@@ -144,7 +144,7 @@ public abstract class HPRestoreItemList extends StaticEntity
 			// the number of times this technique must be used.
 			
 			int hpShort = needed - KoLCharacter.getCurrentHP();
-			int numberToUse = (int) Math.ceil( hpShort / hpPerUse );
+			int numberToUse = (int) Math.ceil( (double) hpShort / (double) hpPerUse );
 
 			if ( TradeableItemDatabase.contains( itemName ) && canUseOtherTechnique )
 				numberToUse = Math.min( numberToUse, itemUsed.getCount( KoLCharacter.getInventory() ) );
